Implement feature as manifest option for kube plugin
