+++ envfrom/plugins/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["cli", "dir", "kube", "yaml"]
+__all__ = ["cli", "dir", "kube"]
